+++ www/js/config.js
@@ -1,0 +1,1 @@
+JOOWAR_SERVER_URL="http://genesis.akikieng.com/joowar-beirut-server";