+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+2015-05-26
+* moving dynamic html added from javascript to angular controller
+
 2015-05-16
 * playing around
 * finished first draft of la madonna